+++ src/main/java/org/socialculture/platform/config/SecurityConfig.java
@@ -1,6 +1,7 @@
 package org.socialculture.platform.config;
 
 import lombok.RequiredArgsConstructor;
+import org.socialculture.platform.member.auth.*;
 import org.socialculture.platform.member.oauth.jwt.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
